requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.1
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
openai==1.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-multipart==0.0.9